--- a/src/lang/type.mli
+++ b/src/lang/type.mli
@@ -75,7 +75,7 @@ type custom = Type_base.custom = ..
 type custom_handler = Type_base.custom_handler = {
   typ : custom;
   copy_with : (t -> t) -> custom -> custom;
-  occur_check : (var -> t -> unit) -> var -> custom -> unit;
+  occur_check : (t -> unit) -> custom -> unit;
   filter_vars : (var list -> t -> var list) -> var list -> custom -> var list;
   repr : (var list -> t -> Repr.t) -> var list -> custom -> Repr.t;
   subtype : (t -> t -> unit) -> custom -> custom -> unit;