--- a/src/lang/types/type_base.ml
+++ b/src/lang/types/type_base.ml
@@ -157,7 +157,7 @@ type custom = ..
 type custom_handler = {
   typ : custom;
   copy_with : (t -> t) -> custom -> custom;
-  occur_check : (var -> t -> unit) -> var -> custom -> unit;
+  occur_check : (t -> unit) -> custom -> unit;
   filter_vars : (var list -> t -> var list) -> var list -> custom -> var list;
   repr : (var list -> t -> R.t) -> var list -> custom -> R.t;
   subtype : (t -> t -> unit) -> custom -> custom -> unit;