At d6c353069abc5636000e6d1c441fcd5c0dd19163:
```
% time ./_build/default/src/bin/liquidsoap.exe ./src/libs/stdlib.liq 'print(1)'
1
No output defined, nothing to do.
zsh: exit 1     ./_build/default/src/bin/liquidsoap.exe ./src/libs/stdlib.liq 'print(1)'
./_build/default/src/bin/liquidsoap.exe ./src/libs/stdlib.liq 'print(1)'  1.81s user 0.06s system 95% cpu 1.959 total
```

At f0ee7d84d68344d35d91b9a1ec644e7110d36a9d:
```
% time ./_build/default/src/bin/liquidsoap.exe ./src/libs/stdlib.liq 'print(1)'
1
No output defined, nothing to do.
zsh: exit 1     ./_build/default/src/bin/liquidsoap.exe ./src/libs/stdlib.liq 'print(1)'
./_build/default/src/bin/liquidsoap.exe ./src/libs/stdlib.liq 'print(1)'  1.42s user 0.03s system 99% cpu 1.454 total
```
About `~20%` increase!

Things I tried:
- [x] Cache types <-> variables mappings. This makes it way worst as neither `WeakHash` or even a specialized `Hashtbl` with type indexes are able to keep up with the sheer amount of stored types and variables.